+++ src/main/java/com/example/vvs/domain/common/s3/S3Config.java
@@ -2,14 +2,15 @@ package com.example.vvs.domain.common.s3;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 
 @Configuration
-public class AwsS3Config {
+public class S3Config {
 
     @Value("${cloud.aws.credentials.access-key}")
     private String accessKey;
@@ -21,15 +22,13 @@ public class AwsS3Config {
     private String region;
 
     @Bean
-    public BasicAWSCredentials basicAWSCredentials() {
-        return new BasicAWSCredentials(accessKey, secretKey);
-    }
+    @Primary
+    public AmazonS3Client amazonS3Client() {
+        BasicAWSCredentials awsCredentials = new BasicAWSCredentials(accessKey, secretKey);
 
-    @Bean
-    public AmazonS3 amazonS3(BasicAWSCredentials basicAWSCredentials) {
-        return AmazonS3ClientBuilder.standard()
+        return (AmazonS3Client) AmazonS3ClientBuilder.standard()
                 .withRegion(region)
-                .withCredentials(new AWSStaticCredentialsProvider(basicAWSCredentials))
+                .withCredentials(new AWSStaticCredentialsProvider(awsCredentials))
                 .build();
     }
 }
