+++ src/main/java/com/example/vvs/domain/common/s3/test/AwsS3Config.java
@@ -1,16 +1,16 @@
-package com.example.vvs.domain.common;
+/*
+package com.example.vvs.domain.common.s3;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
 
 @Configuration
-public class S3Config {
+public class AwsS3Config {
 
     @Value("${cloud.aws.credentials.access-key}")
     private String accessKey;
@@ -22,13 +22,16 @@ public class S3Config {
     private String region;
 
     @Bean
-    @Primary
-    public AmazonS3Client amazonS3Client() {
-        BasicAWSCredentials awsCredentials = new BasicAWSCredentials(accessKey, secretKey);
+    public BasicAWSCredentials basicAWSCredentials() {
+        return new BasicAWSCredentials(accessKey, secretKey);
+    }
 
-        return (AmazonS3Client) AmazonS3ClientBuilder.standard()
+    @Bean
+    public AmazonS3 amazonS3(BasicAWSCredentials basicAWSCredentials) {
+        return AmazonS3ClientBuilder.standard()
                 .withRegion(region)
-                .withCredentials(new AWSStaticCredentialsProvider(awsCredentials))
+                .withCredentials(new AWSStaticCredentialsProvider(basicAWSCredentials))
                 .build();
     }
 }
+*/
