+++ src/main/java/com/example/vvs/domain/common/s3/test/AwsS3Controller.java
@@ -1,4 +1,5 @@
-package com.example.vvs.domain.common.s3;
+/*
+package com.example.vvs.domain.common.s3.test;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -23,3 +24,4 @@ public class AwsS3Controller {
         awsS3Service.remove(awsS3);
     }
 }
+*/
