+++ src/main/java/com/example/vvs/domain/common/s3/test/AwsS3Service.java
@@ -1,10 +1,10 @@
-package com.example.vvs.domain.common.s3;
+/*
+package com.example.vvs.domain.common.s3.test;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.example.vvs.exception.ApiException;
-import com.example.vvs.exception.ErrorHandling;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -82,3 +82,4 @@ public class AwsS3Service {
         amazonS3.deleteObject(bucket, awsS3.getKey());
     }
 }
+*/
